refactor: SceneLabelをsrc/constantsへ移動
Enum定数はmodelsよりもconstantsに配置する方が適切なため移行。
あわせて6ファイルのインポート文を更新。
